pyproject.toml
[project]
name = "youtubevideoassistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.0",
    "langchain-chroma>=0.2.5",
    "langchain-core>=0.3.0",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.31",
    "pytube>=15.0.0",
    "python-dotenv>=1.0.0",
    "youtube-transcript-api<=1.2.2",
]
